feat: create error class and handle exceptions

+++ src/storage/group/groupCreate.ts
@@ -1,12 +1,19 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { GROUP_COLLECTION } from "../storageConfig";
 import { groupsGetAll } from "./groupsGetAll";
+import { AppError } from "@/utils/AppError";
+
 export async function groupCreate(newGroupName: string) {
   try {
     const storedGroups = await groupsGetAll();
 
-    const storage = JSON.stringify([...storedGroups, newGroupName]);
+    const groupAlreadyExists = storedGroups.includes(newGroupName);
 
+    if (groupAlreadyExists) {
+      throw new AppError("storage_appError_groupAlreadyExists");
+    }
+
+    const storage = JSON.stringify([...storedGroups, newGroupName]);
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
   } catch (error) {
     throw error;
